applications: nrf_desktop: switch to 64bit timestamp in eventq
Switch to 64bit timestamp using
k_uptime_get API to calculate timestamp
differences in eventq utility.
We do it in order to prevent overflows.

Jira: NCSDK-34009

+++ applications/nrf_desktop/src/util/hid_eventq.h
@@ -119,7 +119,7 @@ void hid_eventq_reset(struct hid_eventq *q);
  * @param[in] q			HID event queue object.
  * @param[in] min_timestamp	Minimal valid timestamp.
  */
-void hid_eventq_cleanup(struct hid_eventq *q, uint32_t min_timestamp);
+void hid_eventq_cleanup(struct hid_eventq *q, int64_t min_timestamp);
 
 #ifdef __cplusplus
 }
